Initial Commit
This initial commit contains some of the core objects to the app along with all standard REST APIs. These include users (with the ability to login/logout), companies and shares. One can also buy shares from a company, view their shares and create shares for a company. A user's wallet balance may also be deposited to.